A few separate fixes (#92)
A few separate fixes:

* Patching dependencies
* Fixing retries in the restclient
* Tweaking eslint config to work nicely with playwright

I noticed we were getting a load of errors in the test logs and it's because the retry handler didn't have access to the scope where STATUS_CODES was defined. The tests were passing and I think it must just retry anyway if the retry handler throws an error.

```
  console.error
    TypeError: Cannot read properties of undefined (reading 'STATUS_CODES')
        at Request._retryCallback (/Users/andrew.lee/code/hmpps-typescript-lib/packages/rest-client/src/main/RestClient.ts:102:23)
        at Request._retryCallback [as _shouldRetry] (/Users/andrew.lee/code/hmpps-typescript-lib/node_modules/superagent/src/request-base.js:196:29)
```